[#43] refactor : RegisterInterest 관련 컴포넌트 리팩토링 중
