+++ src/components/User/RegisterInterest/const.ts
@@ -1,0 +1,4 @@
+const INIT_INDEX = -1;
+const MAX_SELECT_NUM = 5;
+
+export { INIT_INDEX, MAX_SELECT_NUM };
