+++ src/components/User/RegisterInterest/type.ts
@@ -1,0 +1,19 @@
+import { Dispatch, SetStateAction } from "react";
+
+type TSelectItem = {
+  mainIdx: number;
+  subIdx: number;
+  value?: string;
+};
+type TSelectedInfo = {
+  selectedMainIdx: number;
+  isMax: boolean;
+  items: TSelectItem[];
+};
+
+type TInterestSelectProps = {
+  selectedInfo: TSelectedInfo;
+  setSelectedInfo: Dispatch<SetStateAction<TSelectedInfo>>;
+};
+
+export type { TSelectItem, TSelectedInfo, TInterestSelectProps };
